resolving the issue with authOptions
